backend/pyproject.toml
[project]
name = "meal-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.135.1",
    "firebase-admin>=7.3.0",
    "google-cloud-firestore>=2.25.0",
    "playwright>=1.58.0",
    "pydantic>=2.12.5",
    "sqlalchemy[asyncio]>=2.0.48",
    "uvicorn[standard]>=0.42.0",
]
